+++ tests/test_job_compilers.py
@@ -578,7 +578,60 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
         # Create a timestamp before we start mocking datetime.datetime.
         self.created = datetime.datetime(2018, 7, 6, 11, 52, 33, tzinfo=tz_util.utc)
 
+    def test_chunk_generator(self):
+        from flamenco.job_compilers import blender_render_progressive as brp
+        cg = brp.ChunkGenerator(400, 100)
+        expected = [
+            (1, 10),
+            (11, 49),
+            (50, 133),
+            (134, 222),
+            (223, 311),
+            (312, 400),
+        ]
+        self.assertEqual(expected, list(cg))
+        self.assertEqual(expected, list(cg), 're-iterating "cg" should work')
+
+        cg = brp.ChunkGenerator(30, 3000, uncapped_chunks=3)
+        expected = [
+            (1, 1),
+            (2, 10),
+            (11, 30),
+        ]
+        self.assertEqual(expected, list(cg))
+
     def test_nonexr_job(self):
+        from flamenco.job_compilers import blender_render_progressive
+        from flamenco.exceptions import JobSettingError
+
+        job_doc = JobDocForTesting({
+            '_id': ObjectId(24 * 'f'),
+            '_created': self.created,
+            'settings': {
+                'frames': '1-6',
+                'chunk_size': 2,
+                'render_output': '/render/out/frames-######',
+                'format': 'JPEG',
+                'filepath': '/agent327/scenes/someshot/somefile.blend',
+                'blender_cmd': '/path/to/blender --enable-new-depsgraph',
+                'cycles_sample_count': 30,
+                'cycles_sample_cap': 5,
+            },
+            'job_type': 'blender-render-progressive',
+        })
+        task_manager = mock.Mock()
+        job_manager = mock.Mock()
+        compiler = blender_render_progressive.BlenderRenderProgressive(
+            task_manager=task_manager, job_manager=job_manager)
+
+        try:
+            compiler.compile(job_doc)
+        except JobSettingError as ex:
+            self.assertIn('EXR', str(ex))
+        else:
+            self.fail('expected exception JobSettingError not raised')
+
+    def test_old_style_job(self):
         from flamenco.job_compilers import blender_render_progressive
         from flamenco.exceptions import JobSettingError
 
@@ -602,7 +655,12 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
         compiler = blender_render_progressive.BlenderRenderProgressive(
             task_manager=task_manager, job_manager=job_manager)
 
-        self.assertRaises(JobSettingError, compiler.compile, job_doc)
+        try:
+            compiler.compile(job_doc)
+        except JobSettingError as ex:
+            self.assertIn('Blender Cloud add-on', str(ex))
+        else:
+            self.fail('expected exception JobSettingError not raised')
 
     @mock.patch('datetime.datetime')
     def test_small_job(self, mock_datetime):
@@ -615,13 +673,16 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
             '_created': self.created,
             'settings': {
                 'frames': '1-5',
-                'chunk_size': 2,
+                'chunk_size': 3,
                 'render_output': '/render/out/frames-######',
                 'format': 'EXR',
                 'filepath': '/agent327/scenes/someshot/somefile.blend',
                 'blender_cmd': '/path/to/blender --enable-new-depsgraph',
                 'cycles_sample_count': 30,
-                'cycles_num_chunks': 3,
+
+                # Effectively uncapped so that the number of tasks stays small.
+                # The actual capping is tested in test_chunk_generator() anyway.
+                'cycles_sample_cap': 30,
             },
             'job_type': 'blender-render-progressive',
         })
@@ -632,11 +693,18 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
         mock_now = datetime.datetime.now(tz=tz_util.utc)
         mock_datetime.now.side_effect = [mock_now]
 
-        task_ids = [ObjectId() for _ in range(17)]
+        # We expect:
+        # - 1 destroy-intermediate task
+        # - 2 frame chunks x 3 sample chunks = 6 render tasks
+        # - 4 sample merge tasks
+        # - 1 'publish first chunk' task
+        # so a total of 12 tasks
+        task_ids = [ObjectId() for _ in range(50)]
         task_manager.api_create_task.side_effect = task_ids
 
         compiler = blender_render_progressive.BlenderRenderProgressive(
             task_manager=task_manager, job_manager=job_manager)
+        compiler._uncapped_chunk_count = 3  # Reduce to a testable number of tasks.
         compiler.compile(job_doc)
 
         task_manager.api_create_task.assert_has_calls([
@@ -656,13 +724,12 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
                     filepath='/agent327/scenes/someshot/somefile.blend',
                     format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-######',
-                    frames='1,2',
-                    cycles_num_chunks=3,
-                    cycles_chunk=1,
-                    cycles_samples_from=1,
-                    cycles_samples_to=10)],
-                'render-smpl1-10-frm1,2',
+                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-######',
+                    frames='1..3',
+                    cycles_num_chunks=30,
+                    cycles_chunk_start=1,
+                    cycles_chunk_end=1)],
+                'render-smpl1-1-frm1-3',
                 priority=0,
                 parents=[task_ids[0]],
                 status='under-construction',
@@ -674,31 +741,12 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
                     filepath='/agent327/scenes/someshot/somefile.blend',
                     format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-######',
-                    frames='3,4',
-                    cycles_num_chunks=3,
-                    cycles_chunk=1,
-                    cycles_samples_from=1,
-                    cycles_samples_to=10)],
-                'render-smpl1-10-frm3,4',
-                priority=0,
-                parents=[task_ids[0]],
-                status='under-construction',
-                task_type='blender-render',
-            ),
-            mock.call(  # task 3
-                job_doc,
-                [commands.BlenderRenderProgressive(
-                    blender_cmd='/path/to/blender --enable-new-depsgraph',
-                    filepath='/agent327/scenes/someshot/somefile.blend',
-                    format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-######',
-                    frames='5',
-                    cycles_num_chunks=3,
-                    cycles_chunk=1,
-                    cycles_samples_from=1,
-                    cycles_samples_to=10)],
-                'render-smpl1-10-frm5',
+                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-######',
+                    frames='4,5',
+                    cycles_num_chunks=30,
+                    cycles_chunk_start=1,
+                    cycles_chunk_end=1)],
+                'render-smpl1-1-frm4,5',
                 priority=0,
                 parents=[task_ids[0]],
                 status='under-construction',
@@ -707,87 +755,67 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
 
             # Pre-existing render output dir is moved aside, and intermediate is destroyed.
             # Copy first sample chunk of frames to the output directory.
-            mock.call(  # task 4
+            mock.call(  # task 3
                 job_doc,
                 [
                     commands.MoveOutOfWay(src='/render/out'),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000001.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000001.exr',
                         dest='/render/out/frames-000001.exr',
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000002.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000002.exr',
                         dest='/render/out/frames-000002.exr',
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000003.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000003.exr',
                         dest='/render/out/frames-000003.exr',
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000004.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000004.exr',
                         dest='/render/out/frames-000004.exr',
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000005.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000005.exr',
                         dest='/render/out/frames-000005.exr',
                     ),
                 ],
                 'publish-first-chunk',
-                parents=task_ids[1:4],
+                parents=task_ids[1:3],
                 status='under-construction',
                 task_type='file-management',
             ),
 
             # Second Cycles chunk renders to intermediate directory.
+            mock.call(  # task 4
+                job_doc,
+                [commands.BlenderRenderProgressive(
+                    blender_cmd='/path/to/blender --enable-new-depsgraph',
+                    filepath='/agent327/scenes/someshot/somefile.blend',
+                    format='EXR',
+                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-frm-######',
+                    frames='1..3',
+                    cycles_num_chunks=30,
+                    cycles_chunk_start=2,
+                    cycles_chunk_end=10)],
+                'render-smpl2-10-frm1-3',
+                priority=-10,
+                parents=[task_ids[0]],
+                status='under-construction',
+                task_type='blender-render',
+            ),
             mock.call(  # task 5
                 job_doc,
                 [commands.BlenderRenderProgressive(
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
                     filepath='/agent327/scenes/someshot/somefile.blend',
                     format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-######',
-                    frames='1,2',
-                    cycles_num_chunks=3,
-                    cycles_chunk=2,
-                    cycles_samples_from=11,
-                    cycles_samples_to=20)],
-                'render-smpl11-20-frm1,2',
-                priority=-10,
-                parents=[task_ids[0]],
-                status='under-construction',
-                task_type='blender-render',
-            ),
-            mock.call(  # task 6
-                job_doc,
-                [commands.BlenderRenderProgressive(
-                    blender_cmd='/path/to/blender --enable-new-depsgraph',
-                    filepath='/agent327/scenes/someshot/somefile.blend',
-                    format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-######',
-                    frames='3,4',
-                    cycles_num_chunks=3,
-                    cycles_chunk=2,
-                    cycles_samples_from=11,
-                    cycles_samples_to=20)],
-                'render-smpl11-20-frm3,4',
-                priority=-10,
-                parents=[task_ids[0]],
-                status='under-construction',
-                task_type='blender-render',
-            ),
-            mock.call(  # task 7
-                job_doc,
-                [commands.BlenderRenderProgressive(
-                    blender_cmd='/path/to/blender --enable-new-depsgraph',
-                    filepath='/agent327/scenes/someshot/somefile.blend',
-                    format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-######',
-                    frames='5',
-                    cycles_num_chunks=3,
-                    cycles_chunk=2,
-                    cycles_samples_from=11,
-                    cycles_samples_to=20)],
-                'render-smpl11-20-frm5',
+                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-frm-######',
+                    frames='4,5',
+                    cycles_num_chunks=30,
+                    cycles_chunk_start=2,
+                    cycles_chunk_end=10)],
+                'render-smpl2-10-frm4,5',
                 priority=-10,
                 parents=[task_ids[0]],
                 status='under-construction',
@@ -795,142 +823,112 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
             ),
 
             # First merge pass, outputs to intermediate directory and copies to output dir
-            mock.call(  # task 8
+            mock.call(  # task 6
                 job_doc,
                 [
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000001.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-000001.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000001.exr',
-                        weight1=10,
-                        weight2=10,
+                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000001.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-frm-000001.exr',
+                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000001.exr',
+                        weight1=1,
+                        weight2=9,
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000001.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000001.exr',
                         dest='/render/out/frames-000001.exr',
                     ),
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000002.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-000002.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000002.exr',
-                        weight1=10,
-                        weight2=10,
+                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000002.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-frm-000002.exr',
+                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000002.exr',
+                        weight1=1,
+                        weight2=9,
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000002.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000002.exr',
                         dest='/render/out/frames-000002.exr',
                     ),
-                ],
-                'merge-to-smpl20-frm1,2',
-                parents=[task_ids[4], task_ids[5]],
-                priority=-11,
-                status='under-construction',
-                task_type='exr-merge',
-            ),
-            mock.call(  # task 9
-                job_doc,
-                [
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000003.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-000003.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000003.exr',
-                        weight1=10,
-                        weight2=10,
+                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000003.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-frm-000003.exr',
+                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000003.exr',
+                        weight1=1,
+                        weight2=9,
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000003.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000003.exr',
                         dest='/render/out/frames-000003.exr',
                     ),
-                    commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000004.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-000004.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000004.exr',
-                        weight1=10,
-                        weight2=10,
-                    ),
-                    commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000004.exr',
-                        dest='/render/out/frames-000004.exr',
-                    ),
                 ],
-                'merge-to-smpl20-frm3,4',
-                parents=[task_ids[4], task_ids[6]],
+                'merge-to-smpl10-frm1-3',
+                parents=[task_ids[3], task_ids[4]],
                 priority=-11,
                 status='under-construction',
                 task_type='exr-merge',
             ),
-            mock.call(  # task 10
+            mock.call(  # task 7
                 job_doc,
                 [
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0010-frm-000005.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0020-frm-000005.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000005.exr',
-                        weight1=10,
-                        weight2=10,
+                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000004.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-frm-000004.exr',
+                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000004.exr',
+                        weight1=1,
+                        weight2=9,
                     ),
                     commands.CopyFile(
-                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000005.exr',
+                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000004.exr',
+                        dest='/render/out/frames-000004.exr',
+                    ),
+                    commands.MergeProgressiveRenders(
+                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-frm-000005.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-frm-000005.exr',
+                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000005.exr',
+                        weight1=1,
+                        weight2=9,
+                    ),
+                    commands.CopyFile(
+                        src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000005.exr',
                         dest='/render/out/frames-000005.exr',
                     ),
                 ],
-                'merge-to-smpl20-frm5',
-                parents=[task_ids[4], task_ids[7]],
+                'merge-to-smpl10-frm4,5',
+                parents=[task_ids[3], task_ids[5]],
                 priority=-11,
                 status='under-construction',
                 task_type='exr-merge',
             ),
 
             # Third Cycles chunk renders to intermediate directory.
-            mock.call(  # task 11
+            mock.call(  # task 8
                 job_doc,
                 [commands.BlenderRenderProgressive(
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
                     filepath='/agent327/scenes/someshot/somefile.blend',
                     format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-######',
-                    frames='1,2',
-                    cycles_num_chunks=3,
-                    cycles_chunk=3,
-                    cycles_samples_from=21,
-                    cycles_samples_to=30)],
-                'render-smpl21-30-frm1,2',
+                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-frm-######',
+                    frames='1..3',
+                    cycles_num_chunks=30,
+                    cycles_chunk_start=11,
+                    cycles_chunk_end=30)],
+                'render-smpl11-30-frm1-3',
                 priority=-20,
                 parents=[task_ids[0]],
                 status='under-construction',
                 task_type='blender-render',
             ),
-            mock.call(  # task 12
+            mock.call(  # task 9
                 job_doc,
                 [commands.BlenderRenderProgressive(
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
                     filepath='/agent327/scenes/someshot/somefile.blend',
                     format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-######',
-                    frames='3,4',
-                    cycles_num_chunks=3,
-                    cycles_chunk=3,
-                    cycles_samples_from=21,
-                    cycles_samples_to=30)],
-                'render-smpl21-30-frm3,4',
-                priority=-20,
-                parents=[task_ids[0]],
-                status='under-construction',
-                task_type='blender-render',
-            ),
-            mock.call(  # task 13
-                job_doc,
-                [commands.BlenderRenderProgressive(
-                    blender_cmd='/path/to/blender --enable-new-depsgraph',
-                    filepath='/agent327/scenes/someshot/somefile.blend',
-                    format='EXR',
-                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-######',
-                    frames='5',
-                    cycles_num_chunks=3,
-                    cycles_chunk=3,
-                    cycles_samples_from=21,
-                    cycles_samples_to=30)],
-                'render-smpl21-30-frm5',
+                    render_output='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-frm-######',
+                    frames='4,5',
+                    cycles_num_chunks=30,
+                    cycles_chunk_start=11,
+                    cycles_chunk_end=30)],
+                'render-smpl11-30-frm4,5',
                 priority=-20,
                 parents=[task_ids[0]],
                 status='under-construction',
@@ -940,87 +938,77 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
             # Final merge pass. Could happen directly to the output directory, but to ensure the
             # intermediate directory shows a complete picture (pun intended), we take a similar
             # approach as earlier merge passes.
-            mock.call(  # task 14
+            mock.call(  # task 10
                 job_doc,
                 [
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000001.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-000001.exr',
+                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000001.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-frm-000001.exr',
                         output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000001.exr',
-                        weight1=20,
-                        weight2=10,
+                        weight1=10,
+                        weight2=20,
                     ),
                     commands.CopyFile(
                         src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000001.exr',
                         dest='/render/out/frames-000001.exr',
                     ),
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000002.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-000002.exr',
+                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000002.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-frm-000002.exr',
                         output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000002.exr',
-                        weight1=20,
-                        weight2=10,
+                        weight1=10,
+                        weight2=20,
                     ),
                     commands.CopyFile(
                         src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000002.exr',
                         dest='/render/out/frames-000002.exr',
                     ),
-                ],
-                'merge-to-smpl30-frm1,2',
-                parents=[task_ids[8], task_ids[11]],
-                priority=-21,
-                status='under-construction',
-                task_type='exr-merge',
-            ),
-            mock.call(  # task 15
-                job_doc,
-                [
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000003.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-000003.exr',
+                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000003.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-frm-000003.exr',
                         output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000003.exr',
-                        weight1=20,
-                        weight2=10,
+                        weight1=10,
+                        weight2=20,
                     ),
                     commands.CopyFile(
                         src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000003.exr',
                         dest='/render/out/frames-000003.exr',
                     ),
+                ],
+                'merge-to-smpl30-frm1-3',
+                parents=[task_ids[6], task_ids[8]],
+                priority=-21,
+                status='under-construction',
+                task_type='exr-merge',
+            ),
+            mock.call(  # task 11
+                job_doc,
+                [
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000004.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-000004.exr',
+                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000004.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-frm-000004.exr',
                         output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000004.exr',
-                        weight1=20,
-                        weight2=10,
+                        weight1=10,
+                        weight2=20,
                     ),
                     commands.CopyFile(
                         src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000004.exr',
                         dest='/render/out/frames-000004.exr',
                     ),
-                ],
-                'merge-to-smpl30-frm3,4',
-                parents=[task_ids[9], task_ids[12]],
-                priority=-21,
-                status='under-construction',
-                task_type='exr-merge',
-            ),
-            mock.call(  # task 16
-                job_doc,
-                [
                     commands.MergeProgressiveRenders(
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0020-frm-000005.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0021-0030-frm-000005.exr',
+                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-frm-000005.exr',
+                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-frm-000005.exr',
                         output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000005.exr',
-                        weight1=20,
-                        weight2=10,
+                        weight1=10,
+                        weight2=20,
                     ),
                     commands.CopyFile(
                         src='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-frm-000005.exr',
                         dest='/render/out/frames-000005.exr',
                     ),
                 ],
-                'merge-to-smpl30-frm5',
-                parents=[task_ids[10], task_ids[13]],
+                'merge-to-smpl30-frm4,5',
+                parents=[task_ids[7], task_ids[9]],
                 priority=-21,
                 status='under-construction',
                 task_type='exr-merge',